requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-dotenv==1.0.0
pymongo==4.5.0
langchain==0.0.350
langchain-groq==0.0.1
chromadb==0.4.18
sentence-transformers==2.2.2
Pillow==10.0.1
python-docx==0.8.11
pandas==2.1.3
numpy==1.24.3
Flask-CORS==4.0.0
httpx==0.25.2
requests==2.31.0
werkzeug==2.3.7